Finish first working version of Validated_double_min_max_in_box

+++ chapter_16/exercise/e10/valid_in.cpp
@@ -301,7 +301,7 @@ Validated_double_in_box::Validated_double_in_box(Graph_lib::Point top_left,
                                                  int width,
                                                  int height,
                                                  const string& label,
-                                                 int default_value_as_double,
+                                                 double default_value_as_double,
                                                  Graph_lib::Callback callback_fn)
     : Validated_in_box(top_left,
                        width,
